Enable select all in iframe based webviews
