+++ src/vs/workbench/contrib/webview/browser/webview.ts
@@ -93,6 +93,8 @@ export interface Webview extends IDisposable {
 	hideFind(): void;
 	runFindAction(previous: boolean): void;
 
+	selectAll(): void;
+
 	windowDidDragStart(): void;
 	windowDidDragEnd(): void;
 }
