+++ src/vs/workbench/contrib/webview/browser/webviewCommands.ts
@@ -13,6 +13,7 @@ import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegis
 import { KEYBINDING_CONTEXT_WEBVIEW_FIND_WIDGET_FOCUSED, KEYBINDING_CONTEXT_WEBVIEW_FIND_WIDGET_VISIBLE } from 'vs/workbench/contrib/webview/browser/webview';
 import { WebviewEditor } from 'vs/workbench/contrib/webview/browser/webviewEditor';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
+import { InputFocusedContextKey } from 'vs/platform/contextkey/common/contextkeys';
 
 export class ShowWebViewEditorFindWidgetAction extends Action2 {
 	public static readonly ID = 'editor.action.webvieweditor.showFind';
@@ -97,6 +98,29 @@ export class WebViewEditorFindPreviousCommand extends Action2 {
 		getActiveWebviewEditor(accessor)?.find(true);
 	}
 }
+
+export class SelectAllWebviewEditorCommand extends Action2 {
+	public static readonly ID = 'editor.action.webvieweditor.selectAll';
+	public static readonly LABEL = nls.localize('editor.action.webvieweditor.selectAll', 'Select all');
+
+	constructor(contextKeyExpr: ContextKeyExpr) {
+		const precondition = ContextKeyExpr.and(contextKeyExpr, ContextKeyExpr.not(InputFocusedContextKey));
+		super({
+			id: SelectAllWebviewEditorCommand.ID,
+			title: SelectAllWebviewEditorCommand.LABEL,
+			keybinding: {
+				when: precondition,
+				primary: KeyMod.CtrlCmd | KeyCode.KEY_A,
+				weight: KeybindingWeight.EditorContrib
+			}
+		});
+	}
+
+	public run(accessor: ServicesAccessor, args: any): void {
+		getActiveWebviewEditor(accessor)?.selectAll();
+	}
+}
+
 export class ReloadWebviewAction extends Action {
 	static readonly ID = 'workbench.action.webview.reloadWebviewAction';
 	static readonly LABEL = nls.localize('refreshWebviewLabel', "Reload Webviews");
